move everythin
