--- Assignment/Assignment/SampleData.cs
@@ -1,35 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-
-namespace Assignment;
-
-public class SampleData : ISampleData
-{
-    // 1.
-    public virtual IEnumerable<string> CsvRows => 
-        File.ReadLines("People.csv").Skip(1);
-
-    // 2.
-    public IEnumerable<string> GetUniqueSortedListOfStatesGivenCsvRows() =>
-        CsvRows
-            .Select(row => row.Split(',')[6])
-            .Distinct()                       
-            .OrderBy(state => state);          
-
-    // 3.
-    public string GetAggregateSortedListOfStatesUsingCsvRows()
-        => throw new NotImplementedException();
-
-    // 4.
-    public IEnumerable<IPerson> People => throw new NotImplementedException();
-
-    // 5.
-    public IEnumerable<(string FirstName, string LastName)> FilterByEmailAddress(
-        Predicate<string> filter) => throw new NotImplementedException();
-
-    // 6.
-    public string GetAggregateListOfStatesGivenPeopleCollection(
-        IEnumerable<IPerson> people) => throw new NotImplementedException();
-}
